Fix checkTargets
