+++ main/src/main/scala/sbt/ProjectExtra.scala
@@ -357,31 +357,6 @@ trait ProjectExtra extends Scoped.Syntax:
     def setCond[T](key: AttributeKey[T], vopt: Option[T], attributes: AttributeMap): AttributeMap =
       attributes.setCond(key, vopt)
 
-    private[sbt] def checkTargets(data: Settings[Scope]): Option[String] =
-      val dups = overlappingTargets(allTargets(data))
-      if (dups.isEmpty) None
-      else {
-        val dupStrs = dups map { case (dir, scopes) =>
-          s"${dir.getAbsolutePath}:\n\t${scopes.mkString("\n\t")}"
-        }
-        Some(s"Overlapping output directories:${dupStrs.mkString}")
-      }
-
-    private[this] def overlappingTargets(
-        targets: Seq[(ProjectRef, File)]
-    ): Map[File, Seq[ProjectRef]] =
-      targets.groupBy(_._2).view.filter(_._2.size > 1).mapValues(_.map(_._1)).toMap
-
-    private[this] def allTargets(data: Settings[Scope]): Seq[(ProjectRef, File)] = {
-      import ScopeFilter._
-      val allProjects = ScopeFilter(Make.inAnyProject)
-      val targetAndRef = Def.setting { (Keys.thisProjectRef.value, Keys.target.value) }
-      new SettingKeyAll(Def.optional(targetAndRef)(identity))
-        .all(allProjects)
-        .evaluate(data)
-        .flatMap(x => x)
-    }
-
     private[sbt] def equalKeys(a: ScopedKey[_], b: ScopedKey[_], mask: ScopeMask): Boolean =
       a.key == b.key && Scope.equal(a.scope, b.scope, mask)
